+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVYVQDQWMYKEPH863F7WCB6T"
+        "01KW1REMNAQ4C8KHBM1HG7HKKD"
     )
     r"""The unique identifier of the data source"""