+++ packages/orq-rc/RELEASES.md
@@ -8999,3 +8999,13 @@ Based on:
 - [python v4.11.0-rc.49] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.49] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.49 - packages/orq-rc
+
+## 2026-06-26 10:45:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.785.0 (2.912.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.50] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.50] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.50 - packages/orq-rc